Vari Fix, folder confusion matrix

+++ Company/main/src/app.py
@@ -1,5 +1,5 @@
 # app.py
-
+import os
 import time
 import logging
 import pandas as pd
@@ -17,15 +17,21 @@ def genera_log():
 
 # Configura il sistema di logging
 # logging.basicConfig(filename='logs/company_log.txt', level=logging.INFO)
-print("LOG APPLICATION STARTED")
+print("LOG APPLICATION STARTING...")
 # Esegui il servizio che genera log ogni tot secondi
 input("Please click any keys to start")
 columns = ["event","machine","timestamp","label"]
 df = pd.DataFrame(generate(n=1))
-df.to_csv(
-        "logs/company_log.csv",sep="," , mode="w", header=columns , index=False
-    )
+if os.path.exists("logs/company_log.csv"):
+     df.to_csv(
+            "logs/company_log.csv",sep="," , mode="a", header=0 , index=False
+        )
+else:
+    df.to_csv(
+            "logs/company_log.csv",sep="," , mode="w", header=columns , index=False
+        )
 
+print("Generating logs...")
 while True:
     genera_log()
     time.sleep(1)  # intervallo di 10 secondi
